worktree: symref should be found anywhere in chain
This change fixes the issue where head may be left invalid when
deleting a checked out symref.

Previously, when searching for a shared symref, it fully dereferences
the symref chain before checking the name. This poses problems for
`git branch -d` which will happily delete a checked out symref.

The existing behaviour still exists, but it will now also find any
symref that is in between the starting symref and the first non-symbolic ref.

Concretely, for the following chain
HEAD -> symref1 -> symref2 -> master
previously the function was only able to find master.
With these changes applied, it will be able to find any of the above references

Signed-off-by: Kenneth Cochran <kenneth.cochran101@gmail.com>